+++ packages/experimental/Switch/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui-react-native/switch
 
-This log was last generated on Tue, 03 Jan 2023 23:38:40 GMT and should not be manually modified.
+This log was last generated on Wed, 04 Jan 2023 16:55:53 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.10
+
+Wed, 04 Jan 2023 16:55:53 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.19.8
+- Bump @fluentui-react-native/interactive-hooks to v0.22.7
 
 ## 0.8.9